Move gardena_bluetooth base entity to separate module
